pyproject.toml
[project]
name = "git-commit-mcp-server"
version = "0.1.0"
description = "MCP server for automated Git commit workflows with conventional commit messages"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Git Commit MCP Server Contributors" }
]
dependencies = [
    "fastmcp>=0.1.0",
    "gitpython>=3.1.0",
]

[project.scripts]
git-commit-mcp = "git_commit_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/git_commit_mcp"]
